refactor: add category menu component
