+++ src/components/category-menu/category-menu.styles.scss
@@ -1,0 +1,9 @@
+.categories-container {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(30%, 1fr));
+    /* Auto-fit categories */
+    gap: 15px;
+    width: 95%;
+    padding: 15px;
+    margin: 0 auto;
+}
